V5.1.3.0 release.

+++ sdk/tlkstk/bt/bth/bth.h
@@ -31,7 +31,7 @@
  * Function: BTH Init interface
  * Descript: This interface be used to initial bth layer.
  * Params:
- * Reutrn: Returning TLK_ENONE(0x00) means the send process success.
+ * Return: Returning TLK_ENONE(0x00) means the send process success.
  *         If others value is returned means the send process fail.
 *******************************************************************************/
 int  bth_init(void);
@@ -51,7 +51,7 @@ int bth_sendLeaveSleepCmd(void);
  * Descript: This interface be used to process the event and timer evt, seems 
  *           like a schdluer of System.
  * Params: None.
- * Reutrn: None.
+ * Return: None.
 *******************************************************************************/
 void bth_handler(void);
 
@@ -59,7 +59,7 @@ void bth_handler(void);
  * Function: BTH Destroy interface
  * Descript: This interface be used to release the resource which allocate.
  * Params: @aclhandle[IN]--The acl link handle.
- * Reutrn: None.
+ * Return: None.
 *******************************************************************************/
 void bth_destroy(uint16 aclHandle);
 
@@ -68,7 +68,7 @@ void bth_destroy(uint16 aclHandle);
  * Descript: This interface be used to verify peer device is headset or 
  *           others device.
  * Params: @aclhandle[IN]--The acl link handle.
- * Reutrn: Return true means peer device type headset, false means others .
+ * Return: Return true means peer device type headset, false means others .
 *******************************************************************************/
 bool bth_isHeadset(uint16 aclHandle);
 
@@ -76,7 +76,7 @@ bool bth_isHeadset(uint16 aclHandle);
  * Function: bth_devClassToDevType
  * Descript: This interface be used to reflect from devClass to device type.
  * Params: @devClass[IN]--The device class.
- * Reutrn: The Device type.
+ * Return: The Device type.
 *******************************************************************************/
 uint08 bth_devClassToDevType(uint32 devClass);
 
@@ -85,7 +85,7 @@ uint08 bth_devClassToDevType(uint32 devClass);
  * Descript: This interface be used to set the role of acl link.
  * Params: @btaddr[IN]--The device bt address.
  *         @initRole[IN]--The role of acl link.
- * Reutrn: TLK_ENONE is set sucess, others means failure.
+ * Return: TLK_ENONE is set success, others means failure.
 *******************************************************************************/
 extern int  bth_acl_setInitRole(uint08 btaddr[6], uint08 initRole);
 
@@ -95,17 +95,17 @@ extern int  bth_acl_setInitRole(uint08 btaddr[6], uint08 initRole);
  * Params: @btaddr[IN]--The device bt address.
  *         @devClass[IN]--The device class.
  *         @initRole[IN]--The role of acl link.
- *         @timeout[IN]--The acl setup timeout value
- * Reutrn: TLK_ENONE is set sucess, others means failure.
+ *         @timeout[IN]--The acl setup timeout value. Range:3000~3600000.
+ * Return: TLK_ENONE is set success, others means failure.
 *******************************************************************************/
-extern int  bth_acl_connect(uint08 btaddr[6], uint32 devClass, uint08 initRole, uint16 timeout);
+extern int  bth_acl_connect(uint08 btaddr[6], uint32 devClass, uint08 initRole, uint32 timeout);
 
 /******************************************************************************
  * Function: bth_acl_disconn
  * Descript: This interface be used to tear up the acl link which specify 
  *           by aclhandle.
  * Params: @aclHandle[IN]--The acl link handle.
- * Reutrn: TLK_ENONE is set sucess, others means failure.
+ * Return: TLK_ENONE is set success, others means failure.
 *******************************************************************************/
 extern int  bth_acl_disconn(uint16 aclHandle, uint08 reason);
 
@@ -114,7 +114,7 @@ extern int  bth_acl_disconn(uint16 aclHandle, uint08 reason);
  * Descript: This interface be used to release the acl link resource which
  *           specify by aclhandle.
  * Params: @aclHandle[IN]--The acl link handle.
- * Reutrn: None.
+ * Return: None.
 *******************************************************************************/
 extern void bth_acl_destroy(uint16 aclHandle);
 
@@ -123,7 +123,7 @@ extern void bth_acl_destroy(uint16 aclHandle);
  * Descript: This interface be used to cancel the acl link connect which
  *           specify by bt address.
  * Params: @btaddr[IN]--The peer device address.
- * Reutrn: TLK_ENONE is set sucess, others means failure.
+ * Return: TLK_ENONE is set success, others means failure.
 *******************************************************************************/
 extern int  bth_acl_connectCancel(uint08 btaddr[6]);
 
@@ -132,7 +132,7 @@ extern int  bth_acl_connectCancel(uint08 btaddr[6]);
  * Descript: This interface be used to disconnect the acl link which
  *           specify by bt address.
  * Params: @btaddr[IN]--The peer device address.
- * Reutrn: TLK_ENONE is set sucess, others means failure.
+ * Return: TLK_ENONE is set success, others means failure.
 *******************************************************************************/
 extern int  bth_acl_disconnByAddr(uint08 btaddr[6]);
 
@@ -144,7 +144,7 @@ extern int  bth_acl_disconnByAddr(uint08 btaddr[6]);
  * Params: @aclHandle[IN]--The acl link handle.
  *         @linkType[IN]--it is a scl or sco link type.
  *         @airMode[IN]--The air mode(cvsd or else).
- * Reutrn: TLK_ENONE is set sucess, others means failure.
+ * Return: TLK_ENONE is set success, others means failure.
 *******************************************************************************/
 extern int  bth_sco_connect(uint16 aclHandle, uint16 linkType, uint08 airMode);
 
@@ -154,7 +154,7 @@ extern int  bth_sco_connect(uint16 aclHandle, uint16 linkType, uint08 airMode);
  *           specify by scoHandle.
  * Params: @scoHandle[IN]--The sco link handle.
  *         @reason[IN]--disconnect reason.
- * Reutrn: TLK_ENONE is set sucess, others means failure.
+ * Return: TLK_ENONE is set success, others means failure.
 *******************************************************************************/
 extern int  bth_sco_disconn(uint16 scoHandle, uint08 reason);
 
@@ -162,7 +162,7 @@ extern int  bth_sco_disconn(uint16 scoHandle, uint08 reason);
  * Function: bth_sco_destroy
  * Descript: This interface be used to destroy the SCO link resource.
  * Params: @scoHandle[IN]--The sco link handle.
- * Reutrn: None.
+ * Return: None.
 *******************************************************************************/
 extern void bth_sco_destroy(uint16 scoHandle);
 
@@ -170,7 +170,7 @@ extern void bth_sco_destroy(uint16 scoHandle);
  * Function: bth_sco_destroyByAclHandle
  * Descript: This interface be used to destroy the SCO link resource.
  * Params: @scoHandle[IN]--The sco link handle.
- * Reutrn: None.
+ * Return: None.
 *******************************************************************************/
 extern void bth_sco_destroyByAclHandle(uint16 aclHandle);
 
@@ -180,7 +180,7 @@ extern void bth_sco_destroyByAclHandle(uint16 aclHandle);
  * Params: @scoHandle[IN]--The sco link handle.
  *         @micSpk[IN]--is mic or speaker.
  *         @enable[IN]--enable or disable.
- * Reutrn: TLK_ENONE is set sucess, others means failure.
+ * Return: TLK_ENONE is set success, others means failure.
 *******************************************************************************/
 extern int  bth_sco_sendMute(uint16 scoHandle, uint08 micSpk, uint08 enable);
